move buildOrgUrl to utils
